requirements.txt
Django==5.2.1
django-admin-sortable2==2.2.8
django-tinymce==4.1.0
environs==14.1.1
isort==6.0.1
pillow

